Add Pharo.gitignore
This is the .gitignore that we use for the source management of Pharo.
Prior to version 70, Pharo and its community were not officially using git but one a home made versioning system. 
Since Pharo 70, all the projects are moving to git. 
All the documentation is current heavy changes to reflect the move.